Split recipe listing functionality out from recipeSearchGUI into ScrollableRecipeList. Added search functionality to search screen.
